refactor: Move chart drawing from member to chart
